refactor(ai-agent): isolate launch policy and runtime context
