+++ crates/aionui-ai-agent/src/factory/mod.rs
@@ -1,6 +1,7 @@
 pub mod acp_assembler;
 
 mod acp;
+mod acp_launch_policy;
 pub(crate) mod aionrs;
 mod context;
